requirement.txt
Django==2.2.26
django-sslserver==0.22
djangorestframework==3.13.1
PyMySQL==0.10.0
requests==2.27.1
conda install -c menpo dlib
dlib==19.22.0
face-recognition==1.3.0
opencv-python==4.6.0.66
tensorflow==1.5.0
Keras==2.2.0
scikit-learn==0.23.2
reportlab==3.5.48
redis==4.3.6
PyYAML==6.0
python-dateutil==2.8.2
Pillow==8.4.0
pandas==1.1.5
numpy==1.19.5
joblib==1.1.0
jsonify==0.5
imutils==0.5.4
h5py==2.10.0
xhtml2pdf



   
